+++ hyperactor/src/testing/pingpong.rs
@@ -22,6 +22,7 @@ use crate::Instance;
 use crate::OncePortRef;
 use crate::PortRef;
 use crate::RemoteSpawn;
+use crate::endpoint::Endpoint as _;
 use crate::mailbox::MessageEnvelope;
 use crate::mailbox::Undeliverable;
 use crate::mailbox::UndeliverableMessageError;
