+++ hyperactor_macros/tests/export.rs
@@ -13,6 +13,7 @@ use hyperactor as reference;
 use hyperactor::Actor;
 use hyperactor::Bind;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::Unbind;
 use hyperactor::port::Port;
@@ -131,6 +132,7 @@ hyperactor::behavior!(
 
 #[cfg(test)]
 mod tests {
+    use hyperactor::Endpoint as _;
     use hyperactor::message::ErasedUnbound;
     use hyperactor::message::IndexedErasedUnbound;
     use hyperactor::proc::Proc;
