+++ hyperactor_mesh/src/comm.rs
@@ -25,9 +25,11 @@ use hyperactor::Actor;
 use hyperactor::ActorAddr;
 use hyperactor::ActorRef;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::Instance;
 use hyperactor::PortRef;
+use hyperactor::RemoteEndpoint as _;
 use hyperactor::RemoteMessage;
 use hyperactor::UnboundPort;
 use hyperactor::UnboundPortKind;
