+++ hyperactor_mesh/src/host_mesh/host_agent.rs
@@ -27,6 +27,7 @@ use hyperactor::ActorHandle;
 use hyperactor::ActorRef;
 use hyperactor::Addr;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::HandleClient;
 use hyperactor::Handler;
 use hyperactor::Instance;
@@ -35,6 +36,7 @@ use hyperactor::PortRef;
 use hyperactor::Proc;
 use hyperactor::ProcAddr;
 use hyperactor::RefClient;
+use hyperactor::RemoteEndpoint as _;
 use hyperactor::context;
 use hyperactor::mailbox::MailboxServerHandle;
 use hyperactor_config::Flattrs;
