+++ hyperactor_mesh/src/proc_agent.rs
@@ -25,11 +25,13 @@ use hyperactor::Addr;
 use hyperactor::Bind;
 use hyperactor::Context;
 use hyperactor::Data;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::Instance;
 use hyperactor::PortAddr;
 use hyperactor::PortHandle;
 use hyperactor::PortRef;
+use hyperactor::RemoteEndpoint as _;
 use hyperactor::Unbind;
 use hyperactor::actor::handle_undeliverable_message;
 use hyperactor::actor::remote::Remote;
