+++ hyperactor_mesh/test/hyperactor_mesh_proxy_test.rs
@@ -17,6 +17,7 @@ use clap::Parser;
 use hyperactor as reference;
 use hyperactor::Actor;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::Instance;
 use hyperactor::RemoteSpawn;
