+++ hyperactor_remote/src/token_supervision_tests.rs
@@ -13,6 +13,7 @@ use hyperactor::Actor;
 use hyperactor::ActorAddr;
 use hyperactor::ActorRef;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::Instance;
 use hyperactor::PortHandle;
