+++ monarch_hyperactor/src/code_sync/auto_reload.rs
@@ -13,6 +13,7 @@ use async_trait::async_trait;
 use hyperactor as reference;
 use hyperactor::Actor;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::RemoteSpawn;
 use hyperactor_config::Flattrs;
