+++ monarch_hyperactor/src/local_state_broker.rs
@@ -12,6 +12,7 @@ use async_trait::async_trait;
 use hyperactor::Actor;
 use hyperactor::ActorHandle;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::OncePortHandle;
 use pyo3::prelude::*;
