+++ monarch_rdma/examples/cuda_ping_pong/src/cuda_ping_pong.rs
@@ -63,6 +63,7 @@ use hyperactor::Actor;
 use hyperactor::ActorRef;
 use hyperactor::Bind;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::Handler;
 use hyperactor::Instance;
 use hyperactor::OncePortRef;
