+++ monarch_tensor_worker/src/stream.rs
@@ -24,6 +24,7 @@ use async_trait::async_trait;
 use hyperactor as reference;
 use hyperactor::Actor;
 use hyperactor::Context;
+use hyperactor::Endpoint as _;
 use hyperactor::HandleClient;
 use hyperactor::Handler;
 use hyperactor::Instance;
